+++ src/Utils/ConfigurationKeys.php
@@ -15,7 +15,7 @@ namespace TechDivision\Import\Attribute\Utils;
  *
  * @author    Marcus Döllerer <m.doellerer@techdivision.com>
  * @copyright 2020 TechDivision GmbH <info@techdivision.com>
- * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
+ * @license   https://opensource.org/licenses/MIT
  * @link      https://github.com/techdivision/import-product
  * @link      https://www.techdivision.com
  */