+++ CHANGELOG.md
@@ -11,6 +11,7 @@
     * https://github.com/techdivision/import/issues/181
 * PAC-361: Don't check file system if copy-images defined as false
     * https://github.com/techdivision/import-cli-simple/issues/262
+* PAC-324: Append empty value validator for attribut_option_values
 
 # Version 21.0.0