+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ## Features
 
+* Refactoring deprecated classes. see https://github.com/techdivision/import-cli-simple/blob/master/UPGRADE-4.0.0.md
 * Add #PAC-292: Validate attibute code when importing EAV attributes
 * PAC-96: Use new constands for FileUploadConfiguration
     * https://github.com/techdivision/import/issues/181